refactored CheckoutForm and currently fixing Confirmation bug
