+++ client/components/Order/index.jsx
@@ -1,4 +1,4 @@
 export { default as OrdersList } from './OrdersList';
-export { default as OrderConfirmation } from './OrderConfirmation';
+export { default as Confirmation } from './Confirmation';
 export { default as CheckoutForm } from './CheckoutForm';
 export { default as AddToCart } from './AddToCart';
